ai/01_function_call/requirements.txt
openai>=1.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
tiktoken>=0.7.0
tenacity>=8.3.0
rich>=13.7.0
faiss-cpu>=1.8.0
langchain>=0.2.0
langchain-openai>=0.1.0
numpy>=1.26.0
codebuddy-agent-sdk>=0.1.0
